+++ packages/fluentui/react/src/components/Toolbar/ToolbarMenu.tsx
@@ -30,7 +30,7 @@ import {
 import ToolbarMenuRadioGroup, { ToolbarMenuRadioGroupProps } from './ToolbarMenuRadioGroup';
 import ToolbarMenuDivider from './ToolbarMenuDivider';
 import ToolbarMenuItem, { ToolbarMenuItemProps } from './ToolbarMenuItem';
-import { IconProps } from '../Icon/Icon';
+import { BoxProps } from '../Box/Box';
 
 export type ToolbarMenuItemShorthandKinds = 'divider' | 'item' | 'toggle';
 
@@ -55,7 +55,7 @@ export interface ToolbarMenuProps extends UIComponentProps, ChildrenComponentPro
   submenu?: boolean;
 
   /** Shorthand for the submenu indicator. */
-  submenuIndicator?: ShorthandValue<IconProps>;
+  submenuIndicator?: ShorthandValue<BoxProps>;
 }
 
 export type ToolbarMenuStylesProps = never;
@@ -165,7 +165,7 @@ ToolbarMenu.propTypes = {
   items: customPropTypes.collectionShorthandWithKindProp(['divider', 'item']),
   onItemClick: PropTypes.func,
   submenu: PropTypes.bool,
-  submenuIndicator: customPropTypes.itemShorthandWithoutJSX
+  submenuIndicator: customPropTypes.shorthandAllowingChildren
 };
 ToolbarMenu.handledProps = Object.keys(ToolbarMenu.propTypes) as any;