+++ packages/fluentui/react/src/themes/teams/components/Table/tableCellStyles.ts
@@ -1,10 +1,10 @@
-import { ComponentStyleFunctionParam, ICSSInJSStyle } from '@fluentui/styles';
+import { ComponentSlotStylesPrepared, ICSSInJSStyle } from '@fluentui/styles';
 import { TeamsTableVariables } from './tableVariables';
-import { TableCellProps } from '../../../../components/Table/TableCell';
+import { TableCellStylesProps } from '../../../../components/Table/TableCell';
 import getBorderFocusStyles from '../../getBorderFocusStyles';
 
-export default {
-  root: ({ variables: v, theme: { siteVariables } }: ComponentStyleFunctionParam<TableCellProps, TeamsTableVariables>): ICSSInJSStyle => {
+const tableCellStyles: ComponentSlotStylesPrepared<TableCellStylesProps, TeamsTableVariables> = {
+  root: ({ variables: v, theme: { siteVariables } }): ICSSInJSStyle => {
     const borderFocusStyles = getBorderFocusStyles({
       siteVariables
     });
@@ -25,10 +25,10 @@ export default {
       height: '100%'
     };
   },
-  content: ({ props: { truncateContent } }: ComponentStyleFunctionParam<TableCellProps, TeamsTableVariables>): ICSSInJSStyle => {
+  content: ({ props: p }): ICSSInJSStyle => {
     return {
       alignSelf: 'center',
-      ...(truncateContent && {
+      ...(p.truncateContent && {
         display: 'block',
         overflow: 'hidden',
         textOverflow: 'ellipsis',
@@ -37,3 +37,5 @@ export default {
     };
   }
 };
+
+export default tableCellStyles;