+++ packages/fluentui/react/src/themes/teams/components/Toolbar/toolbarMenuRadioGroupStyles.ts
@@ -1,6 +1,8 @@
-import { ICSSInJSStyle } from '@fluentui/styles';
+import { ComponentSlotStylesPrepared, ICSSInJSStyle } from '@fluentui/styles';
+import { ToolbarMenuRadioGroupStylesProps } from '../../../../components/Toolbar/ToolbarMenuRadioGroup';
+import { ToolbarVariables } from './toolbarVariables';
 
-const toolbarMenuRadioGroupStyles = {
+const toolbarMenuRadioGroupStyles: ComponentSlotStylesPrepared<ToolbarMenuRadioGroupStylesProps, ToolbarVariables> = {
   root: (): ICSSInJSStyle => ({
     padding: 0
   })